Fixes and more schemas
I fixed a few things that weren't working right. The getName() method of MultiPart only returns the name and not the extension of a file, I changed it for getOriginalName.
Turns out forms require setters to edit dtos, who would have thought?. I fixed that.
I also added a second constructor for the personalized exception, in case there is not cause.